Replaced MM functor with M functor
We were told during the lesson that the MM mass gives the combined mass without any topological constraints and that we should thus use the M mass. 

I suppose this translates to the functors? The twiki is not very clear here. In this case, we should be using the M functors instead of the MM functors. I tried to replace them accordingly in this lesson